Commented out code blocks and significant changes in order to remove redundant code and carefully cleanup search features to reduce unneccessary api calls and requests.
