+++ frontend/src/common/utils.test.tsx
@@ -5,8 +5,6 @@ import { message } from 'antd';
 import { Provider, useAtom } from 'jotai';
 import { atomWithStorage } from 'jotai/utils';
 import { vi } from 'vitest';
-import { rawProjectFixture, activeSearchQueryFixture } from '../test/mock/fixtures';
-import { UserSearchQueries, UserSearchQuery } from '../components/Cart/types';
 import { ActiveSearchQuery, RawSearchResult, RawSearchResults } from '../components/Search/types';
 import {
   combineCarts,
@@ -20,7 +18,6 @@ import {
   showError,
   showNotice,
   splitStringByChar,
-  unsavedLocalSearches,
   createSearchRouteURL,
   getStrSizeInKb,
   compressData,
@@ -29,9 +26,6 @@ import {
   getFromLocalStorage,
   cachePagination,
   getCachedPagination,
-  cacheSearchResults,
-  getCachedSearchResults,
-  clearCachedSearchResults,
   showBanner,
   saveBannerText,
   clearDeprecatedStorageKeys,
@@ -39,9 +33,7 @@ import {
   createIntakeEsgfSearch,
   getLastMessageSeen,
   setStartupMessageAsSeen,
-  searchAlreadyExists,
   downloadFileForUser,
-  identifyProblematicFacets,
 } from './utils';
 import { AppPage } from './types';
 import { mockConfig } from '../test/testFunctions';
@@ -315,67 +307,6 @@ describe('Test combineCarts', () => {
   });
   it('returns combined results of 3 items (one duplicate removed)', () => {
     expect(combineCarts(searchResults1, searchResults2).length).toEqual(3);
-  });
-});
-
-describe('Test unsavedLocal searches', () => {
-  const firstResult: UserSearchQuery = {
-    uuid: 'uuid1',
-    user: 'user',
-    project: rawProjectFixture(),
-    projectId: '1',
-    versionType: 'latest',
-    resultType: 'all',
-    minVersionDate: '20200101',
-    maxVersionDate: '20201231',
-    filenameVars: ['var'],
-    activeFacets: { foo: ['option1', 'option2'], baz: ['option1'] },
-    textInputs: ['foo'],
-    url: 'https://localhost/url.com',
-    resultsCount: 200,
-    searchTime: 100000,
-    globusOnly: false,
-  };
-  const secondResult: UserSearchQuery = {
-    uuid: 'uuid2',
-    user: 'user',
-    project: rawProjectFixture(),
-    projectId: '2',
-    versionType: 'latest',
-    resultType: 'all',
-    minVersionDate: '20200101',
-    maxVersionDate: '20201231',
-    filenameVars: ['var'],
-    activeFacets: { foo: ['option1', 'option2'], baz: ['option1'] },
-    textInputs: ['foo'],
-    url: 'https://localhost/url.com',
-    resultsCount: 200,
-    searchTime: 100000,
-    globusOnly: false,
-  };
-  const thirdResult: UserSearchQuery = {
-    uuid: 'uuid3',
-    user: 'user',
-    project: rawProjectFixture(),
-    projectId: '3',
-    versionType: 'latest',
-    resultType: 'all',
-    minVersionDate: '20200101',
-    maxVersionDate: '20201231',
-    filenameVars: ['var'],
-    activeFacets: { foo: ['option1', 'option2'], baz: ['option1'] },
-    textInputs: ['foo'],
-    url: 'https://localhost/url.com',
-    resultsCount: 200,
-    searchTime: 100000,
-    globusOnly: false,
-  };
-
-  const localResults: UserSearchQueries = [firstResult, secondResult];
-  const databaseResults: UserSearchQueries = [secondResult, thirdResult];
-
-  it('returns the first result because it is not currently in database', () => {
-    expect(unsavedLocalSearches(databaseResults, localResults)).toEqual([firstResult]);
   });
 });
 
@@ -818,46 +749,6 @@ describe('Test cachePagination and getCachedPagination', () => {
   });
 });
 
-describe('Test cacheSearchResults, getCachedSearchResults, and clearCachedSearchResults', () => {
-  const results = { response: { docs: [], numFound: 0 } };
-  const pagination = { page: 1, pageSize: 10 };
-  const cachedURL = 'http://test.com';
-
-  afterEach(() => {
-    clearCachedSearchResults();
-  });
-
-  it('caches and retrieves search results', () => {
-    cacheSearchResults(results, pagination, cachedURL);
-    const cached = getCachedSearchResults();
-    expect(cached.cachedURL).toBe(cachedURL);
-    expect(cached.response).toBeDefined();
-  });
-
-  it('clears cached search results', () => {
-    cacheSearchResults(results, pagination, cachedURL);
-    clearCachedSearchResults();
-    expect(getCachedSearchResults()).toEqual({});
-  });
-
-  it('clears cache when expired', () => {
-    cacheSearchResults(results, pagination, cachedURL);
-
-    // Expect the cache to be set
-    expect(tempStorageGetMock('cachedSearchResults')).toBeTruthy();
-
-    // Simulate time passing
-    vi.useFakeTimers();
-    vi.setSystemTime(Date.now() + 60 * 60 * 2000); // Move time forward by 2 hours
-    const cached = getCachedSearchResults();
-    expect(cached).toEqual({});
-
-    // Should also remove from localStorage
-    const cachedItem = tempStorageGetMock('cachedSearchResults');
-    expect(cachedItem).toBeUndefined();
-  });
-});
-
 describe('Test showBanner and saveBannerText', () => {
   beforeEach(() => {
     mockConfig.BANNER_TEXT = 'Test Banner';
@@ -1075,44 +966,6 @@ describe('Test getLastMessageSeen and setStartupMessageAsSeen', () => {
   });
 });
 
-describe('Test searchAlreadyExists', () => {
-  it('returns true if search with same uuid exists', () => {
-    const existingSearches = [
-      {
-        uuid: '123',
-        search: { project: { name: 'CMIP6' }, activeFacets: {} },
-      },
-      {
-        uuid: '456',
-        search: { project: { name: 'CMIP5' }, activeFacets: {} },
-      },
-    ] as unknown as UserSearchQueries;
-
-    const newSearch = {
-      uuid: '123',
-      search: { project: { name: 'CMIP6' }, activeFacets: { activity_id: ['CFMIP'] } },
-    } as unknown as UserSearchQuery;
-
-    expect(searchAlreadyExists(existingSearches, newSearch)).toBe(true);
-  });
-
-  it('returns false if search does not exist', () => {
-    const existingSearches = [
-      {
-        uuid: '123',
-        search: { project: { name: 'CMIP6' }, activeFacets: {} },
-      },
-    ] as unknown as UserSearchQueries;
-
-    const newSearch = {
-      uuid: '789',
-      search: { project: { name: 'E3SM' }, activeFacets: {} },
-    } as unknown as UserSearchQuery;
-
-    expect(searchAlreadyExists(existingSearches, newSearch)).toBe(false);
-  });
-});
-
 describe('Test downloadFileForUser', () => {
   it('creates a download link and triggers download', () => {
     const filename = 'test.txt';
@@ -1148,323 +1001,5 @@ describe('Test downloadFileForUser', () => {
     removeChildSpy.mockRestore();
     clickSpy.mockRestore();
     setAttributeSpy.mockRestore();
-  });
-});
-
-describe('Test identifyProblematicFacets', () => {
-  it('should identify new facets that were not in last successful query', () => {
-    const lastSuccessfulQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1', 'value2'],
-      },
-    });
-
-    const currentQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1', 'value2'],
-        facet2: ['value3'], // New facet
-      },
-    });
-
-    const problematic = identifyProblematicFacets(currentQuery, lastSuccessfulQuery);
-
-    expect(problematic.has('facet2:value3')).toBe(true);
-    expect(problematic.size).toBe(1);
-  });
-
-  it('should identify new values in existing facets', () => {
-    const lastSuccessfulQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1'],
-      },
-    });
-
-    const currentQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1', 'value2'], // Added value2
-      },
-    });
-
-    const problematic = identifyProblematicFacets(currentQuery, lastSuccessfulQuery);
-
-    expect(problematic.has('facet1:value2')).toBe(true);
-    expect(problematic.has('facet1:value1')).toBe(false); // value1 was in successful query
-    expect(problematic.size).toBe(1);
-  });
-
-  it('should identify multiple problematic facets', () => {
-    const lastSuccessfulQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1'],
-      },
-    });
-
-    const currentQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1', 'value2'], // New value
-        facet2: ['value3'], // New facet
-        facet3: ['value4', 'value5'], // New facet with multiple values
-      },
-    });
-
-    const problematic = identifyProblematicFacets(currentQuery, lastSuccessfulQuery);
-
-    expect(problematic.has('facet1:value2')).toBe(true);
-    expect(problematic.has('facet2:value3')).toBe(true);
-    expect(problematic.has('facet3:value4')).toBe(true);
-    expect(problematic.has('facet3:value5')).toBe(true);
-    expect(problematic.size).toBe(4);
-  });
-
-  it('should return empty set when there are no problematic facets', () => {
-    const lastSuccessfulQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1', 'value2'],
-      },
-    });
-
-    const currentQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1', 'value2'], // Same as successful query
-      },
-    });
-
-    const problematic = identifyProblematicFacets(currentQuery, lastSuccessfulQuery);
-
-    expect(problematic.size).toBe(0);
-  });
-
-  it('should return empty set when lastSuccessfulQuery is null', () => {
-    const currentQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1'],
-      },
-    });
-
-    const problematic = identifyProblematicFacets(currentQuery, null);
-
-    expect(problematic.size).toBe(0);
-  });
-
-  it('should handle removed facets (should not mark as problematic)', () => {
-    const lastSuccessfulQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1'],
-        facet2: ['value2'],
-      },
-    });
-
-    const currentQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1'], // facet2 removed
-      },
-    });
-
-    const problematic = identifyProblematicFacets(currentQuery, lastSuccessfulQuery);
-
-    // Removing facets shouldn't be marked as problematic
-    expect(problematic.size).toBe(0);
-  });
-
-  it('should handle empty facets in current query', () => {
-    const lastSuccessfulQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1'],
-      },
-    });
-
-    const currentQuery = activeSearchQueryFixture({
-      activeFacets: {}, // All facets removed
-    });
-
-    const problematic = identifyProblematicFacets(currentQuery, lastSuccessfulQuery);
-
-    expect(problematic.size).toBe(0);
-  });
-
-  it('should handle empty facets in last successful query', () => {
-    const lastSuccessfulQuery = activeSearchQueryFixture({
-      activeFacets: {}, // No facets
-    });
-
-    const currentQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1'], // All facets are new
-      },
-    });
-
-    const problematic = identifyProblematicFacets(currentQuery, lastSuccessfulQuery);
-
-    expect(problematic.has('facet1:value1')).toBe(true);
-    expect(problematic.size).toBe(1);
-  });
-
-  it('should handle facets with array of values', () => {
-    const lastSuccessfulQuery = activeSearchQueryFixture({
-      activeFacets: {
-        experiment: ['historical'],
-      },
-    });
-
-    const currentQuery = activeSearchQueryFixture({
-      activeFacets: {
-        experiment: ['historical', 'rcp85', 'ssp585'], // Added two new values
-      },
-    });
-
-    const problematic = identifyProblematicFacets(currentQuery, lastSuccessfulQuery);
-
-    expect(problematic.has('experiment:rcp85')).toBe(true);
-    expect(problematic.has('experiment:ssp585')).toBe(true);
-    expect(problematic.has('experiment:historical')).toBe(false);
-    expect(problematic.size).toBe(2);
-  });
-
-  it('should be case-sensitive when comparing facet values', () => {
-    const lastSuccessfulQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1'],
-      },
-    });
-
-    const currentQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['Value1'], // Different case
-      },
-    });
-
-    const problematic = identifyProblematicFacets(currentQuery, lastSuccessfulQuery);
-
-    // Case difference should be treated as a new value
-    expect(problematic.has('facet1:Value1')).toBe(true);
-    expect(problematic.size).toBe(1);
-  });
-
-  it('should handle whitespace differences in facet values', () => {
-    const lastSuccessfulQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1'],
-      },
-    });
-
-    const currentQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: [' value1 '], // With extra whitespace
-      },
-    });
-
-    const problematic = identifyProblematicFacets(currentQuery, lastSuccessfulQuery);
-
-    // Whitespace difference should be treated as different
-    expect(problematic.has('facet1: value1 ')).toBe(true);
-    expect(problematic.size).toBe(1);
-  });
-
-  it('should handle special characters in facet names and values', () => {
-    const lastSuccessfulQuery = activeSearchQueryFixture({
-      activeFacets: {
-        'facet-with-dash': ['value.with.dots'],
-      },
-    });
-
-    const currentQuery = activeSearchQueryFixture({
-      activeFacets: {
-        'facet-with-dash': ['value.with.dots', 'value:with:colons'],
-      },
-    });
-
-    const problematic = identifyProblematicFacets(currentQuery, lastSuccessfulQuery);
-
-    expect(problematic.has('facet-with-dash:value:with:colons')).toBe(true);
-    expect(problematic.size).toBe(1);
-  });
-
-  it('should handle complex real-world CMIP6 scenario', () => {
-    const lastSuccessfulQuery = activeSearchQueryFixture({
-      project: rawProjectFixture({ name: 'CMIP6' }),
-      activeFacets: {
-        source_id: ['CESM2', 'GFDL-ESM4'],
-        experiment_id: ['historical'],
-        variable: ['tas'],
-        frequency: ['mon'],
-      },
-    });
-
-    const currentQuery = activeSearchQueryFixture({
-      project: rawProjectFixture({ name: 'CMIP6' }),
-      activeFacets: {
-        source_id: ['CESM2', 'GFDL-ESM4', 'INVALID-MODEL'], // Added invalid model
-        experiment_id: ['historical', 'ssp585'], // Added new experiment
-        variable: ['tas'],
-        frequency: ['mon'],
-        // realm removed, which is fine
-      },
-    });
-
-    const problematic = identifyProblematicFacets(currentQuery, lastSuccessfulQuery);
-
-    expect(problematic.has('source_id:INVALID-MODEL')).toBe(true);
-    expect(problematic.has('experiment_id:ssp585')).toBe(true);
-    expect(problematic.size).toBe(2);
-  });
-
-  it('should handle when current query has subset of successful facets', () => {
-    const lastSuccessfulQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1', 'value2', 'value3'],
-      },
-    });
-
-    const currentQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1', 'value2'], // Subset of successful values
-      },
-    });
-
-    const problematic = identifyProblematicFacets(currentQuery, lastSuccessfulQuery);
-
-    // Using a subset shouldn't be problematic
-    expect(problematic.size).toBe(0);
-  });
-
-  it('should correctly format facet identifiers with colon separator', () => {
-    const lastSuccessfulQuery = activeSearchQueryFixture({
-      activeFacets: {},
-    });
-
-    const currentQuery = activeSearchQueryFixture({
-      activeFacets: {
-        model: ['CESM2'],
-        institution: ['NCAR'],
-      },
-    });
-
-    const problematic = identifyProblematicFacets(currentQuery, lastSuccessfulQuery);
-
-    // Verify the format is "facetKey:facetValue"
-    const problematicArray = Array.from(problematic);
-    expect(problematicArray).toContain('model:CESM2');
-    expect(problematicArray).toContain('institution:NCAR');
-    expect(problematic.size).toBe(2);
-  });
-
-  it('should handle empty arrays for facet values', () => {
-    const lastSuccessfulQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: [],
-      },
-    });
-
-    const currentQuery = activeSearchQueryFixture({
-      activeFacets: {
-        facet1: ['value1'],
-      },
-    });
-
-    const problematic = identifyProblematicFacets(currentQuery, lastSuccessfulQuery);
-
-    // Empty array in last successful means value1 is new
-    expect(problematic.has('facet1:value1')).toBe(true);
-    expect(problematic.size).toBe(1);
   });
 });
